Move error into separate module
